+++ src/app/shared/components/product-list-detail/product-list-detail.ts
@@ -5,6 +5,8 @@ import {RouterLink} from '@angular/router';
 import {TuiFormatNumberPipe, TuiIcon} from '@taiga-ui/core';
 import {ResponsiveBreakpointsService} from '@/services/responsive-breakpoints.service';
 import {FavoritesService} from '@/services/favorites.service';
+import { AccountStore } from '@/account';
+import { AuthenticationOpen } from '@/components/authentication/authentication-open';
 
 @Component({
   selector: 'product-list-detail',
@@ -19,12 +21,18 @@ import {FavoritesService} from '@/services/favorites.service';
 export class ProductListDetail {
   private readonly rbs = inject(ResponsiveBreakpointsService);
   private readonly favoritesService = inject(FavoritesService);
+  private readonly accountStore = inject(AccountStore);
+  private readonly authenticationOpen = inject(AuthenticationOpen);
 
   detail = model.required<ProductListDetailModel>();
   onHome = input(false);
   isMobile = this.rbs.isMobile;
 
   clickFavor() {
+    if (!this.accountStore.account()?.id) {
+      return this.authenticationOpen.openModal();
+    }
+
     if (this.detail().is_favorite) {
       this.favoritesService.delete(this.detail().id).subscribe(() => {
         this.detail.set({...this.detail(), is_favorite: false});
